♻️ Refactor :  약관을 파싱하는 역할와 약관을 렌더하는 역할을 명확히 분리
